+++ src/components/Home/Main/index.js
@@ -9,14 +9,10 @@ export default function HomeIndex() {
     });
   }, []);
   return (
-    <div className="flex flex-col justify-center items-center m-0 p-0 box-border">
-      <div className="max-w-[1280px] w-[100%] ss:px-4 md:px-0 flex justify-center items-center m-auto border-t md:border-0 border-searchBgColor">
-        <div className="w-[100%] h-fit">
-          <MainPageSliders />
-          <CollectionCards />
-          <TypeSection />
-        </div>
-      </div>
+    <div className="w-[100%] h-fit">
+      <MainPageSliders />
+      <CollectionCards />
+      <TypeSection />
     </div>
   );
 }
